requirements.txt
pytest
google-generativeai
pdfminer.six
reportlab
fastapi
uvicorn
celery
redis
